* Several changes to debian build-system: ~ Add debian/get-orig-source  and debian/get-orig-source-from-svn scripts that create orig-tarballs and remove Windows prebuilt binaries ~ Bump Standards-Version to 3.9.1   Stop shipping *.la files ~ Removed doc-packages, because they can not be build directly from svn-sources, a standalone package might come. ~ debian/control   Add cdbs package as Build-Depend   Add libbz2-dev and zlib1g-dev packages as Build-Depends (needed by libhelp_plugin.so)   Add libboost-dev (>=1.40) as Build-Depends (needed for Nassi-Shneiderman-plugin)   Removed Build-Depends for pngnq, wget (was only needed for the removed doc-packages) ~ debian/rules   Switch to CDBS rules system   Add parallel build support   Use lzma compression ~ Added option to keep (or remove) winodows dll's from dist-tarball ~ Some of the changes are based on the work of the ubuntu maintainers, especially Michael Casadevall.
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@6571 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ update_revision.sh
@@ -35,9 +35,18 @@ if [ "x$REV" != "x$OLD_REV" -o ! -r $REV_FILE ]; then
 	echo "m4_define([SVN_DATE], $LCD)" >> $REV_FILE
 
 	# Also change the revision number in debian/changelog for package versioning
-	mv debian/changelog debian/changelog.tmp
-	sed "1 s/(10.05svn[^-)]*/(10.05svn$REV/" < debian/changelog.tmp > debian/changelog
-	rm debian/changelog.tmp
+	if [ -x `which dch` ]; then
+		AKT_REV=`sed -e 's/.*svn\([0-9]*\).*/\1/' -e 'q' < debian/changelog`
+		if [ $REV -gt $AKT_REV ]; then
+			dch -v 10.05svn$REV "New svn revision"
+		fi
+	else
+		echo 'ups'
+
+		mv debian/changelog debian/changelog.tmp
+		sed "1 s/(10.05svn[^-)]*/(10.05svn$REV/" < debian/changelog.tmp > debian/changelog
+		rm debian/changelog.tmp
+	fi
 fi
 
 echo "OLD_REV=$REV" > ./.last_revision
